test: move prompt tests to test directory
To be consistent with `core`, lets move the tests to `test/`.
