+++ packages/prompts/test/note.test.ts
@@ -1,6 +1,6 @@
 import colors from 'picocolors';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest';
-import * as prompts from './index.js';
+import * as prompts from '../src/index.js';
 import { MockReadable, MockWritable } from './test-utils.js';
 
 describe.each(['true', 'false'])('note (isCI = %s)', (isCI) => {
