+++ packages/prompts/test/spinner.test.ts
@@ -1,6 +1,6 @@
 import { EventEmitter } from 'node:stream';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest';
-import * as prompts from './index.js';
+import * as prompts from '../src/index.js';
 import { MockWritable } from './test-utils.js';
 
 describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
